Move layout data to layout.ts  as its flow would be different from normal data in composition.
(Normal data will be populated with pre-order traversal while layout data will have post-order traversal.)
