+++ src/schema/vega.schema.ts
@@ -1,0 +1,8 @@
+export interface VgData {
+  name: string;
+  source?: string;
+  values?: any;
+  format?: any;
+  url?: any;
+  transform?: any;
+}
